power: power_supply: Add property for capping parallel FCC
Add a power supply property to expose the upper limit on the charging
current that the parallel charger can supply.

This is different from POWER_SUPPLY_PROP_CONSTANT_CHARGE_CURRENT_MAX in
that the latter is meant for querying the hardware-defined value, while
this new property is meant to be used to restrict the parallel's share
of FCC to its specified limit much like in SMB1354, for instance.

CRs-Fixed: 2170280
Change-Id: I3a6ed44fbafe0eff4ecc9462bdf9a35f34811d8c
Signed-off-by: Guru Das Srinagesh <gurus@codeaurora.org>
Signed-off-by: Angga Satya Putra <anggasp@hotmail.com>