ANDROID: sched/fair: Cleanup and simplify cpu_util_freq()
Now that we have a proper aggregation of PELT signals in cpu_util_freq()
let's refactor that function to make it more streamling and trust the
compiler in doing the possible optmizations at compile time.

Do that by using unsigned long long for utilization, to cover the case
of WALT where we have u64_t for utilization maths.

Change-Id: Ia1d7705aad7163dd3e1a36dc71358c8aede45a2e
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Angga Satya Putra <anggasp@hotmail.com>